// packages/playground/src/api/mock/files-data.ts
import { IFile } from '@/contexts/files-context'

export const FILES: IFile[] = [
  {
    id: '0ff8cf7c-fca3-45df-a275-e079d15ba90e', // generated by crypto.randomUUID()
    name: 'file1',
    content: `
USE sp500insight;
SELECT sector, industry, COUNT(*) AS companies
FROM companies c
WHERE c.stock_symbol IN (SELECT stock_symbol FROM index_compositions WHERE index_symbol = "SP500")
GROUP BY sector, industry
ORDER BY sector, companies DESC LIMIT 10;
`
  },
  {
    id: '838e098a-752e-445d-bccb-6262c4529378',
    name: 'file2',
    content: `
/* Enter "USE {database};" to start exploring your data.
Type "-- your question" followed by the "ENTER" key to try out AI-generated SQL queries.*/
USE sp500insight;
SELECT sector, industry, COUNT(*) AS companies
FROM companies c
WHERE c.stock_symbol IN (SELECT stock_symbol FROM index_compositions WHERE index_symbol = "SP500")
GROUP BY sector, industry
ORDER BY sector, companies DESC LIMIT 10;
`
  },
  {
    id: 'f0e3f7d4-1a1c-4c0e-9e9b-6f5b7e7d8d4d',
    name: 'file3',
    content: `select aaa from bbb;`
  },
  {
    id: '63f6794d-6158-424f-8cf4-7d03b1b9f48c',
    name: 'file4',
    content: `select ccc from ddd;`
  },
  {
    id: 'd7692e17-b97d-4342-a6b7-a5e6e4d50021',
    name: 'file5',
    content: `select eee from fff;`
  },
  {
    id: '64c80ed3-f7b4-4959-9520-ad92706aa815',
    name: 'file6',
    content: `select ggg from hhh;`
  },
  {
    id: '64c80ed3-f7b4-4959-9520-ad92706aa816',
    name: 'file7',
    content: `USE game;

SELECT g.*
FROM game_genre gg
LEFT JOIN games g ON g.app_id = gg.app_id
WHERE gg.genre_id  = 9
ORDER BY g.estimated_owners DESC;`
  },
  {
    id: '64c80ed3-f7b4-4959-9520-ad92706aa817',
    name: 'file8',
    content: ``
  }
]